Cleaned up publish checks; removed skip filter from prepare
