+++ scripts/release/publish-commands/validate-tags.js
@@ -7,16 +7,14 @@ const {join} = require('path');
 const theme = require('../theme');
 
 const run = async ({cwd, packages, skipPackages, tags}) => {
-  const abitraryPackageName = packages.find(
-    packageName => !skipPackages.includes(packageName)
-  );
-
   // Prevent a canary release from ever being published as @latest
+  // All canaries share a version number, so it's okay to check any of them.
+  const arbitraryPackageName = packages[0];
   const packageJSONPath = join(
     cwd,
     'build',
     'node_modules',
-    abitraryPackageName,
+    arbitraryPackageName,
     'package.json'
   );
   const {version} = await readJson(packageJSONPath);
@@ -24,6 +22,13 @@ const run = async ({cwd, packages, skipPackages, tags}) => {
     if (tags.includes('latest')) {
       console.log(
         theme`{error Canary release} {version ${version}} {error cannot be tagged as} {tag latest}`
+      );
+      process.exit(1);
+    }
+  } else {
+    if (tags.includes('canary')) {
+      console.log(
+        theme`{error Stable release} {version ${version}} {error cannot be tagged as} {tag canary}`
       );
       process.exit(1);
     }
