+++ scripts/release/publish.js
@@ -3,9 +3,11 @@
 'use strict';
 
 const {join} = require('path');
-const {getPublicPackages, filterPackages, handleError} = require('./utils');
+const {getPublicPackages, handleError} = require('./utils');
+const theme = require('./theme');
 
 const checkNPMPermissions = require('./publish-commands/check-npm-permissions');
+const confirmSkippedPackages = require('./publish-commands/confirm-skipped-packages');
 const confirmVersionAndTags = require('./publish-commands/confirm-version-and-tags');
 const downloadErrorCodesFromCI = require('./publish-commands/download-error-codes-from-ci');
 const parseParams = require('./publish-commands/parse-params');
@@ -22,10 +24,22 @@ const run = async () => {
     params.cwd = join(__dirname, '..', '..');
     params.packages = await getPublicPackages();
 
-    // Remove any skipped packages
-    filterPackages(params.packages, params.skipPackages);
+    // Pre-filter any skipped packages to simplify the following commands.
+    // As part of doing this we can also validate that none of the skipped packages were misspelled.
+    params.skipPackages.forEach(packageName => {
+      const index = params.packages.indexOf(packageName);
+      if (index < 0) {
+        console.log(
+          theme`Invalid skip package {package ${packageName}} specified.`
+        );
+        process.exit(1);
+      } else {
+        params.packages.splice(index, 1);
+      }
+    });
 
     await validateTags(params);
+    await confirmSkippedPackages(params);
     await confirmVersionAndTags(params);
     await validateSkipPackages(params);
     await checkNPMPermissions(params);
